tests/fixtures/inventory-agent/requirements.txt
langchain==0.3.1
langchain-openai==0.2.1
langchain-core==0.3.5
pinecone-client==3.1.0
openai==1.12.0
pydantic==2.5.3
